backend/requirements_v2.txt
fastapi==0.115.5
uvicorn==0.32.1
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
cryptography==43.0.3
python-dotenv==1.0.0
pyotp==2.9.0
qrcode[pil]==7.4.2
slowapi==0.1.9
alembic==1.17.1

